+++ skale/contracts/ima/mainnet/__init__.py
@@ -1,0 +1,24 @@
+# flake8: noqa
+
+
+from skale.contracts.base_contract import BaseContract, transaction_method
+
+from skale.contracts.ima.mainnet.message_proxy_for_mainnet import MessageProxyForMainnet
+from skale.contracts.ima.mainnet.linker import Linker
+from skale.contracts.ima.mainnet.community_pool import CommunityPool
+from skale.contracts.ima.mainnet.deposit_box_eth import DepositBoxEth
+from skale.contracts.ima.mainnet.deposit_box_erc20 import DepositBoxERC20
+from skale.contracts.ima.mainnet.deposit_box_erc721 import DepositBoxERC721
+from skale.contracts.ima.mainnet.deposit_box_erc721_wmt import DepositBoxERC721WithMetadata
+from skale.contracts.ima.mainnet.deposit_box_erc1155 import DepositBoxERC1155
+
+__all__ = [
+    'MessageProxyForMainnet',
+    'Linker',
+    'CommunityPool',
+    'DepositBoxEth',
+    'DepositBoxERC20',
+    'DepositBoxERC721',
+    'DepositBoxERC1155',
+    'DepositBoxERC721WithMetadata'
+]
