+++ skale/contracts/ima/schain/__init__.py
@@ -1,0 +1,23 @@
+# flake8: noqa
+
+from skale.contracts.manager.contract_manager import ContractManager
+from skale.contracts.base_contract import BaseContract, transaction_method
+
+from skale.contracts.ima.schain.message_proxy_for_schain import MessageProxyForSchain
+#from skale.contracts.ima_schain.token_manager_linker import TokenManagerLinker
+from skale.contracts.ima.schain.token_manager_eth import TokenManagerETH
+from skale.contracts.ima.schain.token_manager_erc20 import TokenManagerERC20
+from skale.contracts.ima.schain.token_manager_erc721 import TokenManagerERC721
+from skale.contracts.ima.schain.token_manager_erc721_wmt import TokenManagerERC721WithMetadata
+from skale.contracts.ima.schain.token_manager_erc1155 import TokenManagerERC1155
+from skale.contracts.ima.schain.community_locker import CommunityLocker
+
+__all__ = [
+    'MessageProxyForSchain',
+    'TokenManagerETH',
+    'TokenManagerERC20',
+    'TokenManagerERC721',
+    'TokenManagerERC721WithMetadata',
+    'TokenManagerERC1155',
+    'CommunityLocker'
+]
