--- skale/contracts/ima_schain/token_manager_linker.py
@@ -1,7 +1,0 @@
-from skale.contracts.base_contract import BaseContract, transaction_method
-from skale.transactions.result import TxRes
-
-
-
-class TokenManagerLinker(BaseContract):
-    #TODO
